requirements.txt
beautifulsoup4==4.6.0
blinker==1.3
boto==2.38.0
chardet==2.3.0
cloud-init==0.7.9
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
EasyProcess==0.2.3
google-compute-engine==2.3.6
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
lxml==3.8.0
MarkupSafe==0.23
oauthlib==1.0.3
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
PyVirtualDisplay==0.2.1
PyYAML==3.11
requests==2.9.1
selenium==3.4.3
six==1.10.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
